+++ src/js/components/ListTracks/ListTracks.js
@@ -223,7 +223,9 @@ const ListEntry = React.memo(
           )}
 
           {optionShowStarRatings && (
-            <div className={style.userRating}>{entry.userRating && <StarRating rating={entry.userRating} />}</div>
+            <div className={style.userRating}>
+              <StarRating type="track" ratingKey={entry.trackId} rating={entry.userRating} editable />
+            </div>
           )}
 
           <div className={style.duration}>{durationToStringShort(entry.duration)}</div>