+++ src/js/pages/ArtistGenreDetail.js
@@ -6,7 +6,7 @@ import { useEffect } from 'react';
 import { useSelector } from 'react-redux';
 import { useParams } from 'react-router-dom';
 
-import { ListCards, Loading, StarRating, TitleHeading } from 'js/components';
+import { ListCards, Loading, TitleHeading } from 'js/components';
 import * as plex from 'js/services/plex';
 
 // ======================================================================
@@ -16,8 +16,6 @@ import * as plex from 'js/services/plex';
 const ArtistGenreDetail = () => {
   const { genreId, libraryId } = useParams();
 
-  const optionShowStarRatings = useSelector(({ sessionModel }) => sessionModel.optionShowStarRatings);
-
   const allArtistGenres = useSelector(({ appModel }) => appModel.allArtistGenres);
   const currentArtistGenre = allArtistGenres?.filter((genre) => genre.genreId === genreId)[0];
 
@@ -26,7 +24,6 @@ const ArtistGenreDetail = () => {
 
   const genreThumb = currentArtistGenre?.thumb;
   const genreTitle = currentArtistGenre?.title;
-  const genreRating = currentArtistGenre?.userRating;
 
   useEffect(() => {
     plex.getAllArtistGenres();
@@ -39,7 +36,6 @@ const ArtistGenreDetail = () => {
         <TitleHeading
           thumb={genreThumb}
           title={genreTitle}
-          detail={optionShowStarRatings && genreRating && <StarRating rating={genreRating} size={13} inline />}
           subtitle={
             currentArtistGenreItems ? (
               currentArtistGenreItems?.length + ' Artist' + (currentArtistGenreItems?.length !== 1 ? 's' : '')